requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
django-cors-headers>=4.3
python-decouple>=3.8